refactor some class in activity screen
